Make alphagov-deployment files available to app's deployment scripts
Change jenkins.sh to clone alphagov-deployment and copy the files to
upload into a `secrets` subdirectory of the app's directory in this repo.

We will also need to make the shared file upload steps look in the
`secrets` directory, but we think that putting the files into a separate
directory is cleaner than putting them directly into the app's
directory, which runs the risk of overwriting existing files.

Mobbed by @alexmuller @brenetic @emmabeynon @klssmith @jennyd