Don't run migrations by default for efg-training
We've never deployed efg-training before - the list of the migrations
in the EFG app is long and doesn't complete successfully against an
empty database, which makes Capistrano rollback the deploy and remove
the application code.

I want to get a copy of the app deployed so that I can run a rake
task to load the database from `schema.rb`, at which point all future
deploys should be set to run migrations.