Add configuration for govuk-content-schemas
This is needed on backend machines such that the Publishing API can
validate the payloads it receives using the schemas.

Previously the govuk-content-schemas deployment was managed separately,
and while it is not a service with a running process, its current use
via the Publishing API is equivalent to a dependant
application (e.g. the content-store), its just that the interface is
currently over the filesystem, and not over HTTP.