Change search-admin deploy config
search-admin is moving to the new deployment pipeline.

This commit removes the deployment steps where files containing secrets are copied from GHE, overwriting placeholders in the public app repository. These files are included in the app repo as of https://github.com/alphagov/search-admin/pull/71.

It also loads and runs the `govuk_admin_template` task, which sets the environment indicators in the app header.