Only one publishing-api server class has db role
By both of these having a db role we end up running migrations on 2
servers simultaneously which causes a "Cannot run migrations because
another migration process is currently running." error.

This doesn't feel an ideal fix as both classes of server will still be
defined with the first item as a primary, which may not be expected
behaviour.