Run hard_restart command from restart
If we run hard_restart we have the problem that deploy:restart after
hooks don't get called.

It feels like it's easier to fix this problem with a "deploy:restart"
that gets called whatever restart than it is to set up
"deploy:hard_restart" since "deploy:restart" is such a common hook
point.