Switch Slack channel and post on deploy start
We are experimenting with retiring the badger and release calendar and
replacing them with a Slack-based deployment system. This will make
Capistrano post in #govuk-deploy on both start and success of staging
and production deployments.

Also added the version tag being deployed, and an emoji to help
distinguish the environments.