Exclude regression test artefacts
The aim of this commit is to avoid syncing the test/artefacts directory
from the smart answers project.

This directory contains regression test artefacts that are redundant in
production.

These regression test artefacts can be found under a deep directory
structure and we believe this slows down the deployment of smart
answers.

This commit adds an exclude test/artefacts directory option to the
rsync, this will in effect avoid deploying the artefacts directory and
hopefully reduce the taken to deploy smart answers.