Draft deployment bugfix
The set servers script depended on the hostname to work out whether a
server was a draft machine. In Amazon this is redundant because we do
not set hostnames.

Update the script so that we just use the variable we pass to collect
the hostnames. This should work in AWS, and should not break current
deployments.