Only use URI auth if set
Previously the `setup_http_request` method always set the user and
password for HTTP requests even if these were set to empty strings or
nil. When accessing resources in S3 this causes a 400 bad request
response as S3 does not support this method of authorisation.

This only sets up basic auth if the user and password are set.