Enable asset precompilation for Asset Manager
We're going to want to use Rails asset pipeline assets for the
`thumbnail-placeholder.png` image which we're going to move across
from Whitehall.

I've set the `assets_prefix` to 'asset-manager' to match the value in
alphagov/asset-manager#199 and to conform with the approach taken by
other apps, e.g. static, so that we'll be able to route requests via
the Nginx configuration for the `assets-origin` host [1].

Closes alpahgov/asset-manager#201.

[1]:
https://github.com/alphagov/govuk-puppet/blob/994d5424c7e4e28e903039f9f603e8d345b4f01d/hieradata/common.yaml#L1545-L1566