Copy router artefact to deployed-to-environment
In AWS we're using this branch to guess what version of the app to
deploy.

For Router this is not possible because we're just pulling an artefact,
rather than a Github repository.

Copy the released artefact to the name of the thing we're trying to
deploy.

Also update govuk_crawler_worker to use the same methods.