Ignore host key verification
There is a file `/etc/ssh/ssh_known_hosts` that is managed by puppet and
contains the ssh host keys for all hosts on the system. Unfortunately,
setting the capistrano ssh option `global_known_hosts_file` does not use
this file and will still append hosts to the jenkins users known hosts
file, which is not managed by puppet.

In AWS, when a new machine comes up, capistrano will without asking for
verification add this machines host key into the jenkins users known
hosts file. However, if a new machine comes up with an ip address that
is already in the user known hosts file, the host key will not match and
this will cause a deployment error.

Although it feels wrong to ignore the host key, currently when a new
machine comes up it is silently added to the jenkins users ssh
known_hosts file, so if a rogue host on the network came up we would
deploy to it anyway.

For reference the net ssh options capistrano uses are here -
https://net-ssh.github.io/ssh/v2/api/classes/Net/SSH.html#M000002