Add ability to tag existing docker images remotely
We need the ability to tag an existing docker image (eg
`govuk/whitehall:release_12345`) as deployed to an environment similar
to how we do git tagging. Rather that set up docker on the deployment
Jenkins nodes and have them pull, tag then push, we can retrieve the
image manifest (a list of image layers) for the source tag then send it
straight back under the new tag.

This uses the Registry API specified at
https://docs.docker.com/registry/spec/api/#put-manifest. We use v2 of
the manifest schema as it’s much simpler than v1, and can be sent back
verbatim without modification or signature.

The Docker Hub requires a Bearer token, so we need to authenticate with
a Docker ID over Basic Auth first.