Move error handling order for DockerTagPusher
As errors are returned from the Docker registry API with a content type
of application/json they were raising the less descriptive "Remote image
not in correct format" error than the server error one. I have also
modified that to have the status code in the message and dropped the
word "Server" since they could well be client errors.