Task to git tag a release for an artefact release
Repos which are deployed via artefact such as router lack the consistent
deployed-to-* branches that are present on all our non-artefact
branches. This does a clone of the repo for that artefact and tags it.

As far as I could tell there wasn't a way to do this without a git
clone.