Fix router deploy for branch
Turns out that passing a branch name is not ideal as this only works
when the branch is actually a tag. When it is just a branch this fails.

Instead we can use the query_revision method to get a commit sha hash
and use that for the clone instead.