Add restart publishing-api bulk import consumer
Content Performance Manager is subscribed to changes in Publishing API
via a consumer of a Rabbit MQ queue. It is necessary then to restart the
consumer each time the application is restarted.

The queue has a specific routing key, so messages come through to the
just this consumer when we run a bulk import.