Send SHA & user to Release app
This sends additional info the the release app:

- the SHA of what we're actually committing
- the user name and email of the person deploying

This will allow us to show the user on a deploy page, and have better
insight to what we actually deployed. This is especially useful in
cases where branches are deployed by name, which can obviously change.

Release app hasn't been updated to persist this data, that will be done
in a future PR.