Reduces the number of releases we keep
Each release costs disk space and can vary from a couple of Megabytes to
several hundred. The purpose of storing releases on the machine is to
enable fast rollback without having to re-deploy the app. Having 2
releases should be sufficient to preserve this facility and means we can
avoid resizing the disks, as they are currently nearing capacity.