Add deployment configuration for the Content Data API
The Content Performance Manager application is being renamed to the
Content Data API to better reflect it's current function.

The configuration added in this commit relates to the new resources
for the Content Data API. It's basically a copy of the Content
Performance Manager configuration.

Once the migration from the Content Performance Manager to the Content
Data API has been completed, the Content Performance Manager
configuration can be removed.