add licensify apps to capistrano
Licensify is composed of 3 apps: licensify, licensify-admin and
licensify-feed. licensify app is deployed to the licensing_frontend
servers only. licensify-admin and licensify-feed apps are deployed
to the licensing_backend servers only

These apps were deployed using the deprecated alpha_deployment.
This PR will transfer the app deployments to this current repo.

Changes to the deployment process are:
1. each app can be deployed on its own rather than all 3 together
2. since there is only one repo hosting the 3 licensify apps,
   new tag names are used for the `deployed-to-<environment>`
   where the app name is appended in the format:
   `<app-name>-deployed-to-<environment>`
3. the deployment process here does not transfer any secrets
   on the licensing machines, these are set using
   govuk-puppet with its encrypted hiera