Warm finder-frontend cache on deploy
Previously we'd refresh the registry cache on all the boxes.  I don't think
we need to necessarily do that.  We just need to make sure there is something
already there before the app starts accepting requests.  If we're deploying
a new version of finder-frontend then caches will already be populated.

If we need to refresh all the registries, then we can use fabric scripts
and run the `registries:cache_refresh` rake task as before.

Caches are still refreshed on a schedule defined in finder-frontend.