Add a parameter to set_servers to install on only one machine
A new environment variable "TARGET_MACHINES" is now passed to the
capistrano job. If this variable is not "all", then
capistrano deploys to the servers mentioned in the variable.
Otherwise, the deploy process remains as before, i.e.
all servers in a class are deployed to.

"TARGET_MACHINES" format is a string of comma-delimited
FQDN of servers to deploy to or the special keyword
"all".