pass application name to release app
During the migration of licensify to AWS, govuk-app-deployment will be
used to individually deploy the 3 licensify apps in the single repo:
https://github.com/alphagov/licensify since they are hosted on different
machine classes. This is in contrast with the current method of
deploying licensify where a single deployment across multiple machine
classes is performed using the deprecated alphagov-deployment.

Unfortunately, govuk-app-deployment will not be able to record correctly
the 3 individual app deployments with the release app because the
release app uses the repo name to identify an app and the 3 apps share
the same repo.

Hence, we change the notification to the release app in this PR so that
the application name is also sent to the release app to identify which
app to record the new deployment under.

To maintain backward compability with the previous way of identifying an
app by repo name in the release app, the application name will only be
used by the release app if the additional boolean parameter
`application_by_name` is set to true when the notification is made.

The corresponding changes to the release app is done in
https://github.com/alphagov/release/pull/470