Run Mapit using Python 3
This changes the deploy script to installed a Python 3 virtual
environment and run Mapit inside that.

I've changed the directory name of the virtual environment from `venv`
to `venv3` because it's not easy to upgrade the version of Python
inside a virtual enviroment in place. This way we can leave the old
virtual environment working while we switch over in case we have to
rollback for whatever reason.