Recover for 'bundle clean' when deployment fails
Previously an issue with Bundler (see below) caused git dependencies to
be deleted when 'bundle clean' is run. While the files are normally
replaced as part of the subsequent 'bundle install', a failed deployment
(before the 'bundle install' step) will lead to an issue where 'bundle
clean' fails next time round. In the longterm we expect to upgrade to a
patched version of Bundler. This adds a short-term fix to ensure we only
attempt 'bundle clean' when the local gems are all there.