Use the correct node class when deploying to specific nodes.
Fixes deployment of Whitehall when new machine instances come up.

This was broken owing to a bug introduced in #325 when we added support
for deploying to specific hosts. When deploying to a specific instance of
`whitehall_frontend`, we would iterate over the two server_classes for
Whitehall (`whitehall_backend` and `whitehall_frontend`) and each time
around the loop we would run the deployment task on the list of machines
specified in TARGET_MACHINES. This resulted in running a deployment of
`whitehall_frontend` on the specified machine, followed by a deployment
of `whitehall_backend` on the same machine.