Run mapit migrations in correct virtual environment
The previous behaviour meant that it was looking for a virtual
environment called `venv` which no longer exists (it's now called
`venv3` since running on Python 3). This wasn't a problem for our
existing machines because the old virtual environments were kept
around, but on new machines the deploy process was failing.

```
05:11:18   * executing "cd /data/apps/mapit/releases/20200312051058 && govuk_setenv mapit /data/apps/mapit/shared/venv/bin/python manage.py migrate --noinput --settings=project.settings"
05:11:18     servers: ["ip-10-13-4-212.eu-west-1.compute.internal"]
05:11:18     [ip-10-13-4-212.eu-west-1.compute.internal] executing command
05:11:18 *** [err :: ip-10-13-4-212.eu-west-1.compute.internal] envdir: fatal: unable to run /data/apps/mapit/shared/venv/bin/python: file does not exist
```