Bump rubocop-govuk and autofix failures
The rules in rubocop-govuk have changed a bit recently!

One config change that we want is to ignore the ./tmp directory because
this is where gems are cached in Github actions builds.

Rather that just add that to the local config, I thought it better to lean
in the change and get the latest version of the gem with the config upstream.