Use govuk_unicorn_reload
govuk_unicorn_reload is a wrapper script for the command to reload the
unicorn-herder process. This resolves a problem we currently have where
after `initctl reload $app` is run we don't know which version of the
app will be used in any smoke tests as both are running.
govuk_unicorn_reload monitors the switching of these processes and
terminates once it is complete.

This also removes the constraint of running this command only on a
single host at a time as this delays deploys somewhat by multiplying the
wait sequentially. This does provide a slightly higher risk that a bad
deploy could break all versions of the app, but that is basically the
situation already.