Downgrade capistrano to 2.13.5
Later versions have changes to 'deploy:assets:precompile' which cause
the current deployment process to break with a 'No such file or directory'
error for '/data/apps/<app>/shared/assets/manifest*' when trying to handle
the differences between sprockets 2 and 3. This version still has the
removal of the scanf require which was the reason to upgrade initially.