Set real_revision to the deployed branch
The real_revision configuration variable is used internally by
Capistrano by write out the REVISION file when deploying an app:

https://github.com/capistrano/capistrano/blob/08a82f3618425eae7539e2bbd34a87e35bac2800/lib/capistrano/recipes/deploy/strategy/remote.rb#L43-L47

https://github.com/capistrano/capistrano/blob/08a82f3618425eae7539e2bbd34a87e35bac2800/lib/capistrano/recipes/deploy/strategy/base.rb#L83-L87

By setting the configuration variable manually ourselves, we change it
from the default (a Git SHA of the tag) to a more human-readable string.

This value is currently used by Sentry to display the release of the
app, and using the Git SHA makes it harder to quickly see at a glance
the release of the app:

https://github.com/getsentry/sentry-ruby/blob/ace176cfceaefc91e29364a1c4180a67337e8ab0/sentry-raven/lib/raven/configuration.rb#L481-L494

Instead we can show the full name of the tag or branch which was
deployed (for example `release_xxx`) which makes it a lot easier to
cross reference that with any recent deploys in the Release app to know
when a specific issue was introduced in Sentry.

By changing this value, Capistrano will also use this when performing
other Git related commands, however this shouldn't be a problem as
anywhere that Git accepts the Git SHA, it should accept the name of the
branch or tag.