Write revision file with human readable string
By default Capistrano writes out the Git SHA of the release deployed to this file, but we'd prefer it to be a more human readable string. By changing it to the `TAG` that was deployed this should be the case.

Specifically, this helps us navigate issues in Sentry as the more human readable string will be used to associate issues with the release it was introduced in, allowing us to better correlate problems with a release.

This was attempted before in dd7f387cdc4094a98f9ae501d512ed50d8f975d5 but it caused issue with branch deploys.