Simplify Whitehall asset compilation
Unlike every other GOV.UK app, Whitehall is using a precompilation
approach of installing assets on the machine doing the deploy and then
rsyncing them to each server.

This approach doesn't work for an app with yarn dependencies as yarn
isn't available on the deploying host. Rather than going on a big yak
shave to enable that I felt it would be best to make Whitehall deploy
assets the same way all the other apps work - with just a minor extra
step for the shared_mustache compilation.

There's a chance this may perform slightly worse than the previous
approach, but this is vastly simpler and no longer an infrastructure
snowflake.