Port bundle:install task directly into ruby recipe
We rely on an old deprecated way to run bundle install, using a
capistrano file directly from bundler (😮). This has been deprecated for
a long time (2017 [1]) but we can't migrate to the new way [2] as that
requires Capistrano 3 and we've dug ourselves deep into a Capistrano 2
hole.

Anyway the approach used to bundle install in the old deprecated gem was
out-of-date with the world of Bundler 2.1 and produces gem install
errors when used with Ruby 3.1:

```
14:42:44 *** [err :: ip-10-1-6-251.eu-west-1.compute.internal] Following files may not be writable, so sudo is needed:
14:42:44 *** [err :: ip-10-1-6-251.eu-west-1.compute.internal] /usr/lib/rbenv/versions/3.1.2/bin
14:42:44 *** [err :: ip-10-1-6-251.eu-west-1.compute.internal] /usr/lib/rbenv/versions/3.1.2/lib/ruby/gems/3.1.0
14:42:44 *** [err :: ip-10-1-6-251.eu-west-1.compute.internal] /usr/lib/rbenv/versions/3.1.2/lib/ruby/gems/3.1.0/build_info
14:42:44 *** [err :: ip-10-1-6-251.eu-west-1.compute.internal] /usr/lib/rbenv/versions/3.1.2/lib/ruby/gems/3.1.0/cache
14:42:44 *** [err :: ip-10-1-6-251.eu-west-1.compute.internal] /usr/lib/rbenv/versions/3.1.2/lib/ruby/gems/3.1.0/doc
14:42:44 *** [err :: ip-10-1-6-251.eu-west-1.compute.internal] /usr/lib/rbenv/versions/3.1.2/lib/ruby/gems/3.1.0/extensions
14:42:44 *** [err :: ip-10-1-6-251.eu-west-1.compute.internal] /usr/lib/rbenv/versions/3.1.2/lib/ruby/gems/3.1.0/gems
14:42:44 *** [err :: ip-10-1-6-251.eu-west-1.compute.internal] /usr/lib/rbenv/versions/3.1.2/lib/ruby/gems/3.1.0/plugins
14:42:44 *** [err :: ip-10-1-6-251.eu-west-1.compute.internal] /usr/lib/rbenv/versions/3.1.2/lib/ruby/gems/3.1.0/specifications
14:42:44 *** [err :: ip-10-1-6-251.eu-west-1.compute.internal] sudo: no tty present and no askpass program specified
```

This updates the commands sent to bundler to use bundle config rather
than options to bundler - those will be removed from Bundler 3.

I've kept compatibility with the existing configuration argument used
`bundle_without`. I couldn't find any of the other config options [3]
used.

I found I had to cd on each command, doing a single cd step before the
actions failed for me. There is a Capistrano within option, however
that arrived in Capistrano 3.

This requires all the applications to be using Bundler 2.1 and this
would be selected by their version of Ruby. I've been through all the
apps and they all seem to be 2.1 Bundler and Ruby 2.7 so I think we're
good. I didn't really want to add an extra conditional step.

[1]: https://github.com/rubygems/rubygems/commit/9af2ece4d8793497cf0aa2b150e9f32639652388
[2]: https://github.com/capistrano/bundler
[3]: https://github.com/rubygems/rubygems/blob/0497485acba3c09b476ee3ef61b84c099b965643/bundler/lib/bundler/deployment.rb#L39-L45