No longer fetch the rake task
Since removing bundler/capistrano from the previous commit this
argument has started failing by using a rake version without a prefix of
bundle exec. This fixes that by getting rid of the argument and calling
it directly.

I assume the problem occurs because something else sets a rake fetch
somewhere so the fallback argument isn't used. This wasn't a problem
before because bundler/capistrano set it. [1]

[1]: https://github.com/rubygems/rubygems/blob/0497485acba3c09b476ee3ef61b84c099b965643/bundler/lib/bundler/capistrano.rb#L21