Remove docker tagging from deployments
This functionality was used to tag Docker images with tags like
"deployed-to-production" at the point of them being deployed to an
environment.

This functionality was only used by the publishing-e2e-tests [1]
utility which used the tags to determine the current production versions
of applications. Since that utility is being retired we no longer have a
use (at least one I know of) for this so can remove it.

[1]: https://github.com/alphagov/publishing-e2e-tests